sg: Split up commands in separate files
Nothing changed here except that I extracted the code for each `sg`
subcommand into its own file.

That should make it easier for others to contribute because it's easier
to see where the code for a command lives and it's easier to see what's
required to add a new command.
